+++ packages/aeria-sdk/src/index.ts
@@ -1,6 +1,5 @@
 export { topLevel as Aeria } from './topLevel'
 export * from '@sonata-api/common'
 export * from './topLevel'
-
-// @ts-ignore
-export * from '.aeria-sdk'
+export * from './runtime'
+export * from './storage'