+++ packages/aeria-sdk/src/utils.ts
@@ -1,0 +1,12 @@
+import type { InstanceConfig } from './types'
+
+export const apiUrl = (config: InstanceConfig) => {
+  if( typeof config.apiUrl === 'string' ) {
+    return config.apiUrl
+  }
+
+  return process.env.NODE_ENV === 'production'
+    ? config.apiUrl.production
+    : config.apiUrl.development
+}
+