+++ packages/common/src/string.ts
@@ -1,4 +1,0 @@
-export const capitalize = function(target: string) {
-  return target.charAt(0).toUpperCase() + target.slice(1)
-}
-