+++ packages/aeria-sdk/CHANGELOG.md
@@ -1,5 +1,26 @@
 # aeria-sdk
 
+## 0.0.6
+
+### Patch Changes
+
+- Aeria SDK and types improvements
+- Updated dependencies
+  - @sonata-api/common@0.1.13
+  - @sonata-api/types@0.1.18
+
+## 0.0.5
+
+### Patch Changes
+
+- Aeria SDK storage
+
+## 0.0.4
+
+### Patch Changes
+
+- Write .aeria-sdk in resolved node_modules
+
 ## 0.0.3
 
 ### Patch Changes