+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sonata-api/types
 
+## 0.1.18
+
+### Patch Changes
+
+- Aeria SDK and types improvements
+
 ## 0.1.17
 
 ### Patch Changes